+++ Source/Plugin.BLE.Abstractions/Utils/TaskBuilder.cs
@@ -8,7 +8,7 @@ namespace Plugin.BLE.Abstractions.Utils
     {
         public static Task<TReturn> FromEvent<TReturn, TEventHandler>(
             Action execute,
-            Func<Action<TReturn>, TEventHandler> getCompleteHandler,
+            Func<Action<TReturn>, Action<Exception>, TEventHandler> getCompleteHandler,
             Action<TEventHandler> subscribeComplete,
             Action<TEventHandler> unsubscribeComplete,
             CancellationToken token = default(CancellationToken))
@@ -25,7 +25,7 @@ namespace Plugin.BLE.Abstractions.Utils
 
         public static async Task<TReturn> FromEvent<TReturn, TEventHandler, TRejectHandler>(
             Action execute,
-            Func<Action<TReturn>, TEventHandler> getCompleteHandler,
+            Func<Action<TReturn>, Action<Exception>, TEventHandler> getCompleteHandler,
             Action<TEventHandler> subscribeComplete,
             Action<TEventHandler> unsubscribeComplete,
             Func<Action<Exception>, TRejectHandler> getRejectHandler,
@@ -35,9 +35,10 @@ namespace Plugin.BLE.Abstractions.Utils
         {
             var tcs = new TaskCompletionSource<TReturn>();
             Action<TReturn> complete = args => tcs.TrySetResult(args);
+            Action<Exception> completeException = ex => tcs.TrySetException(ex);
             Action<Exception> reject = ex => tcs.TrySetException(ex);
 
-            var handler = getCompleteHandler(complete);
+            var handler = getCompleteHandler(complete, completeException);
             var rejectHandler = getRejectHandler(reject);
 
             try