+++ Source/Plugin.BLE.Android/Descriptor.cs
@@ -1,17 +1,79 @@
 using System;
+using System.Threading.Tasks;
 using Android.Bluetooth;
 using Plugin.BLE.Abstractions;
+using Plugin.BLE.Abstractions.Utils;
+using Plugin.BLE.Android.CallbackEventArgs;
 
 namespace Plugin.BLE.Android
 {
     public class Descriptor : DescriptorBase
     {
         private readonly BluetoothGattDescriptor _nativeDescriptor;
+        private readonly BluetoothGatt _gatt;
+        private readonly IGattCallback _gattCallback;
+
         public override Guid Id => Guid.ParseExact(_nativeDescriptor.Uuid.ToString(), "d");
 
-        public Descriptor(BluetoothGattDescriptor nativeDescriptor)
+        public override byte[] Value => _nativeDescriptor.GetValue();
+
+        public Descriptor(BluetoothGattDescriptor nativeDescriptor, BluetoothGatt gatt, IGattCallback gattCallback)
         {
+            _gattCallback = gattCallback;
+            _gatt = gatt;
+
             _nativeDescriptor = nativeDescriptor;
+        }
+
+        protected override Task WriteNativeAsync(byte[] data)
+        {
+            return TaskBuilder.FromEvent<bool, EventHandler<DescriptorCallbackEventArgs>>(
+               execute: () => InternalWrite(data),
+               getCompleteHandler: (complete, reject) => ((sender, args) =>
+                  {
+                      if (args.Descriptor.Uuid == _nativeDescriptor.Uuid)
+                      {
+                          if (args.Exception != null)
+                              reject(args.Exception);
+                          else
+                              complete(true);
+                      }
+                  }),
+               subscribeComplete: handler => _gattCallback.DescriptorValueWritten += handler,
+               unsubscribeComplete: handler => _gattCallback.DescriptorValueWritten -= handler
+           );
+        }
+
+        private void InternalWrite(byte[] data)
+        {
+            if (!_nativeDescriptor.SetValue(data))
+                throw new Exception("GATT: SET descriptor value failed");
+
+            if (!_gatt.WriteDescriptor(_nativeDescriptor))
+                throw new Exception("GATT: WRITE descriptor value failed");
+        }
+
+        protected override async Task<byte[]> ReadNativeAsync()
+        {
+            return await TaskBuilder.FromEvent<byte[], EventHandler<DescriptorCallbackEventArgs>>(
+               execute: ReadInternal,
+               getCompleteHandler: (complete, reject) => ((sender, args) =>
+                  {
+                      if (args.Descriptor.Uuid == _nativeDescriptor.Uuid)
+                      {
+                          complete(args.Descriptor.GetValue());
+                      }
+                  }),
+                subscribeComplete: handler => _gattCallback.DescriptorValueRead += handler,
+              unsubscribeComplete: handler => _gattCallback.DescriptorValueRead -= handler
+           );
+        }
+
+        void ReadInternal()
+        {
+            if (!_gatt.ReadDescriptor(_nativeDescriptor))
+                throw new Exception("GATT: read characteristic FALSE");
+
         }
     }
 }